Removes E prototype folder and merges with main files
Originally it was thought that the best approach would be to create a
folder and make changes to the prototype there - but this caused issues
with the school lookup script. Renaming the page didn't fix the issue.

Speaking with service design, it was decided we don't need to keep the
prototype as it is and we can make the changes to existing files.