Change the check answers page to show all questions
The check answers page should show all the questions the user has
answered. It checks to see if the question input has a value, and will
then show it if it does.

- Changed the values to inputs to be human readable, Yes instead of
true, No instead of false

- Rewrite the check-answers page so it uses HTML instead of nunjucks to
render the table of answers so that we only show a question of the user
answered it