Add in student loan questions after Verify
We identified that we needed to identify student loan amount and plan
on this journey so we can pay any additional tax. I added in the
questions from the student loans company > what is my student loan plan
so we can help users identify the plan and save that plan for them.
Copy for questions will need improving.

Housekeeping on index page to add in Beta 2 so we can show iteration
when we come to review the service.