Change the contact method to text or email
Previously we allowed users to specify an email address and mobile phone
number so that they could be contacted using both menthods. In UR
it appears that most users would choose email so they could dig it out
again. Our hypothesis that we are testing is that they don't need or
care about multiple contact methods.