Amended Verify Requirements before verify starts
On the eligible confirmed screen we inform users that they appear to be
eligible for the payment and now they need go through verify. We need
to update the requirements to reflect the new LOA2 journey that has been
implemented. We also need to present this information consistently on
this page and the start page.